% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dapp_servers.R
\name{dapp_servers}
\alias{dapp_servers}
\title{Fetch version information}
\usage{
dapp_servers(url, ...)
}
\arguments{
\item{url}{A URL for an OPeNDAP server. Default: see \code{\link{durl}}}

\item{...}{Further args passed on to \code{\link[httr]{GET}} (must
be named parameters)}
}
\description{
NOT READY YET
}
\examples{
\dontrun{
dapp_servers()
}
}

