% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dapp_version}
\alias{dapp_version}
\title{Fetch version information}
\usage{
dapp_version(url, ...)
}
\description{
Fetch version information
}
\examples{
\dontrun{
dapp_version(durl())
dapp_version('http://oceandata.sci.gsfc.nasa.gov/opendap/')
}
}

