% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dapper-package.r
\docType{package}
\name{dapper-package}
\alias{dapper}
\alias{dapper-package}
\title{General purpose R client for ERDDAP servers}
\description{
General purpose R client for ERDDAP servers
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

