% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dappr-package.r
\docType{package}
\name{dappr-package}
\alias{dappr-package}
\alias{dappr}
\title{General purpose R client for ERDDAP servers}
\description{
General purpose R client for ERDDAP servers
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
