% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dds.R
\name{dds}
\alias{dds}
\title{Get dds information on an OPeNDAP dataset.}
\usage{
dds(id, url = durl(), ...)
}
\arguments{
\item{id}{Dataset id}

\item{url}{A URL for an OPeNDAP server. Default: see \code{\link{durl}}}

\item{...}{Further args passed on to \code{\link[httr]{GET}} (must be named parameters)}
}
\value{
Prints a data.frame of metadata for a dataset
}
\description{
Get dds information on an OPeNDAP dataset.
}
\examples{
\dontrun{
dds("MERRA_MONTHLY/MAIMNXINT.5.2.0/1980/MERRA100.prod.assim.instM_2d_int_Nx.198004.hdf")
dds("MERRA_MONTHLY/MSTMNXMLD.5.2.0/2011/MERRA300.prod.simul.tavgM_2d_mld_Nx.201105.hdf")

# NASA's ocean color data
url <- 'http://oceandata.sci.gsfc.nasa.gov/opendap/'
dds('MODISA/L3SMI/2015/006/A2015006.L3m_DAY_CHL_chl_ocx_4km.nc', url)
}
}

