% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ddx.R
\name{ddx}
\alias{ddx}
\title{Get dds information on an OPeNDAP dataset.}
\usage{
ddx(id, url = durl(), ...)
}
\arguments{
\item{id}{Dataset id}

\item{url}{A URL for an ERDDAP server. Default: \url{http://upwell.pfeg.noaa.gov/erddap/}}

\item{...}{Further args passed on to \code{\link[httr]{GET}} (must be a named parameter)}
}
\value{
Prints a summary of the data on return, but you can index to various information.
}
\description{
FIXME - messy, messy parsing.
}
\examples{
\dontrun{
ddx(id = "MERRA_MONTHLY/MSTMNXMLD.5.2.0/2015/MERRA300.prod.simul.tavgM_2d_mld_Nx.201507.hdf")
}
}
\author{
Scott Chamberlain <myrmecocystus@gmail.com>
}

