% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_info.R
\name{as_info}
\alias{as_info}
\title{Coerce to info object}
\usage{
as_info(x, url, ...)
}
\arguments{
\item{x}{input}

\item{url}{A URL for an OPeNDAP server. Default: see \code{\link{durl}}}

\item{...}{Further args passed on to \code{\link[httr]{GET}} (must
be named parameters)}
}
\description{
Coerce to info object
}
