% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/od_list.R
\name{od_list}
\alias{od_list}
\title{List datasets for an OPeNDAP server}
\usage{
od_list(..., url = durl(), callopts = list())
}
\arguments{
\item{...}{Any number of paths to drill down into an OPeNDAP server}

\item{url}{A URL for an OPeNDAP server. Default: see \code{\link{durl}}}

\item{callopts}{Curl options}
}
\description{
List datasets for an OPeNDAP server
}
\examples{
\dontrun{
od_list("MERRA_MONTHLY")
od_list("MERRA_MONTHLY", "MSTMNXMLD.5.2.0")
od_list("MERRA_MONTHLY", "MATMNXLND.5.2.0")
od_list("MERRA_MONTHLY", "MSTMNXMLD.5.2.0", "2000")
od_list("MERRA_MONTHLY", "MSTMNXMLD.5.2.0", "2015")
}
}

