% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.R
\name{od_search}
\alias{od_search}
\title{Search for datasets}
\usage{
od_search(..., url = durl(), callopts = list())
}
\arguments{
\item{...}{Any number of paths to drill down into an OPeNDAP server}

\item{url}{A URL for an OPeNDAP server. Default: see \code{\link{durl}}}

\item{callopts}{Curl options}
}
\description{
Search for datasets
}
\examples{
\dontrun{
od_search("MERRA_MONTHLY")
od_search("MERRA_MONTHLY", "MSTMNXMLD.5.2.0")
od_search("MERRA_MONTHLY", "MATMNXLND.5.2.0")
od_search("MERRA_MONTHLY", "MSTMNXMLD.5.2.0", "2000")
od_search("MERRA_MONTHLY", "MSTMNXMLD.5.2.0", "2015")
}
}

