% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{dapp_version}
\alias{dapp_version}
\title{Fetch version information}
\usage{
dapp_version(url, ...)
}
\arguments{
\item{url}{A URL for an OPeNDAP server. Default: see \code{\link{durl}}}

\item{...}{Further args passed on to \code{\link[httr]{GET}} (must
be named parameters)}
}
\description{
Fetch version information
}
\examples{
\dontrun{
dapp_version(durl())
dapp_version('http://oceandata.sci.gsfc.nasa.gov/opendap/')
}
}
