% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch}
\alias{fetch}
\title{Get files}
\usage{
fetch(x, ..., fields = "all", stride = 1, fmt = "nc", url = durl(),
  store = disk(), read = TRUE, callopts = list())
}
\arguments{
\item{...}{Further args passed on to \code{\link[httr]{GET}} (must
be named parameters)}

\item{url}{A URL for an OPeNDAP server. Default: see \code{\link{durl}}}

\item{id}{Dataset id}
}
\value{
Prints a summary of the data on return, but you can index to various information.
}
\description{
Get files
}
\examples{
\dontrun{
fetch(id = "MERRA_MONTHLY/MSTMNXMLD.5.2.0/2015/MERRA300.prod.simul.tavgM_2d_mld_Nx.201507.hdf")

# NASA's ocean color data
url <- 'http://oceandata.sci.gsfc.nasa.gov/opendap/'
fetch('MODISA/L3SMI/2015/006/A2015006.L3m_DAY_IOP_a_412_giop_4km.nc', url)

x <- "MODISA/L3SMI/2015/006/A2015006.L3m_DAY_IOP_aph_unc_443_giop_9km.nc.nc4"
}
}
\author{
Scott Chamberlain <myrmecocystus@gmail.com>
}

